+++ src/track.h
@@ -544,7 +544,7 @@ typedef struct {
   s32 TOW_ms;                /**< TOW in ms. */
   s32 TOW_ms_prev;           /**< previous TOW in ms. */
   s32 TOW_residual_ns;       /**< Residual to TOW_ms [ns] */
-  u32 sample_count;          /**< Total num samples channel has tracked for. */
+  u64 sample_count;          /**< Total num samples channel has tracked for. */
   double code_phase_prompt;  /**< Prompt code phase in chips. */
   double code_phase_rate;    /**< Code phase rate in chips/s. */
   double carrier_phase;      /**< Carrier phase in cycles. */
@@ -913,7 +913,7 @@ bool tracker_channel_available(tracker_channel_id_t id,
 bool tracker_channel_init(tracker_channel_id_t id,
                           const me_gnss_signal_t mesid,
                           u16 glo_orbit_slot,
-                          u32 ref_sample_count,
+                          u64 ref_sample_count,
                           double code_phase,
                           float carrier_freq,
                           u32 chips_to_correlate,
@@ -923,14 +923,14 @@ bool tracker_channel_disable(tracker_channel_id_t id);
 /* Tracking parameters interface. */
 
 void tracking_channel_measurement_get(
-    double ref_tc,
+    u64 ref_tc,
     const tracking_channel_info_t *info,
     const tracking_channel_freq_info_t *freq_info,
     const tracking_channel_time_info_t *time_info,
     const tracking_channel_misc_info_t *misc_info,
     channel_measurement_t *meas);
 
-bool tracking_channel_calc_pseudorange(double ref_tc,
+bool tracking_channel_calc_pseudorange(u64 ref_tc,
                                        const channel_measurement_t *meas,
                                        double *raw_pseudorange);
